Backend/requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.5
beautifulsoup4==4.9.1
boto3==1.14.12
botocore==1.17.12
cachetools==2.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.15
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-graphql-auth==0.3.9
django-graphql-jwt==0.3.0
django-graphql-social-auth==0.1.4
django-heroku==0.3.1
django-markdown==0.8.4
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.15.2
filelock==3.0.12
future==0.16.0
gkeepapi==0.11.2
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-vision==0.42.0
googleapis-common-protos==1.51.0
gpsoauth==0.4.1
graphene==2.1.8
graphene-django==2.9.0
graphene-file-upload==1.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.27.2
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.11.3
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
jmespath==0.9.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.2.1
mccabe==0.6.1
mongoengine==0.19.1
msgpack-python==0.5.6
oauth2client==4.1.3
oauthlib==3.1.0
pathtools==0.1.2
Pillow==7.0.0
promise==2.3
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycparser==2.20
pycryptodomex==3.9.7
pygame==1.9.6
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
pylru==1.2.0
pymongo==3.10.1
PyOpenGL==3.1.5
pyparsing==2.4.6
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.12.0
python-resize-image==1.1.19
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
robobrowser==0.5.3
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
singledispatch==3.4.0.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==3.3.3
social-auth-storage-mongoengine==1.0.1
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.47.0
termcolor==1.1.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.8
virtualenv==20.0.13
virtualenvwrapper-win==1.2.6
watchdog==0.10.2
Werkzeug==0.14.1
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
